move truth modules into main path
